backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pydantic>=2.9.0
python-dotenv==1.0.0
httpx>=0.28.1
resend>=2.0.0
